requirements.txt
PyQt6==6.8.1
requests==2.32.3
cinemagoer==2023.5.1
duckduckgo_search==7.5.0
langchain==0.3.20
langchain-openai==0.3.8
openai==1.65.4
pydantic==2.10.6
python-dotenv==1.0.1
